:art: Abstract Primary/Secondary logic from Symptom
