+++ labsys/admissions/mixins.py
@@ -34,3 +34,28 @@ class DatedEvent(db.Model):
         return '<{}[{}]: {}>'.format(self.__class__.__name__,
                                      self.id,
                                      self.occurred)
+
+
+class PrimarySecondaryEntity(db.Model):
+    '''
+    A named entity that is classified as primary or secondary
+    '''
+    __abstract__ = True
+    id = db.Column(db.Integer, primary_key=True)
+    # Attributes
+    name = db.Column(db.String(64))
+    primary = db.Column(db.Boolean)
+
+    @classmethod
+    def get_primary(cls):
+        return cls.query.filter(
+            cls.primary is True).order_by(asc(cls.name)).all()
+
+    @classmethod
+    def get_secondary(cls):
+        return cls.query.filter(
+            cls.primary is False).order_by(asc(cls.name)).all()
+
+    def __repr__(self):
+        return '<{}[{}]: {}, Primary: {}>'.format(
+            self.__class__.__name__, self.id, self.name, self.primary)
